chore(dark-mode): add the onModeChange event on layout component
